Build 8 and 10 bits versions of rcn_inter.c

+++ libovvc/bitdepth.h
@@ -2,7 +2,7 @@
 #define BITDEPTH_H
 
 #if BITDEPTH > 8
-typedef uint8_t OVSample;
+typedef uint16_t OVSample;
 #else
 typedef uint8_t OVSample;
 #endif
